# Conflicts:
#	SteelBot/DataProviders/SubProviders/AuditLogProvider.cs
#	SteelBot/DiscordModules/AuditLog/AuditLogSlashCommands.cs

+++ SteelBot/DataProviders/SubProviders/SelfRoleProvider.cs
@@ -1,9 +1,8 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
-using Sentry;
 using SteelBot.Database;
 using SteelBot.Database.Models;
-using SteelBot.Helpers.Sentry;
+using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Linq;
@@ -13,17 +12,15 @@ namespace SteelBot.DataProviders.SubProviders;
 
 public class SelfRolesProvider
 {
-    private readonly ILogger<SelfRolesProvider> _logger;
     private readonly IDbContextFactory<SteelBotContext> _dbContextFactory;
-    private readonly IHub _sentry;
+    private readonly ILogger<SelfRolesProvider> _logger;
 
     private readonly ConcurrentDictionary<ulong, ConcurrentDictionary<ulong, SelfRole>> _selfRolesByGuildAndId;
 
-    public SelfRolesProvider(ILogger<SelfRolesProvider> logger, IDbContextFactory<SteelBotContext> contextFactory, IHub sentry)
+    public SelfRolesProvider(ILogger<SelfRolesProvider> logger, IDbContextFactory<SteelBotContext> contextFactory)
     {
         _logger = logger;
         _dbContextFactory = contextFactory;
-        _sentry = sentry;
 
         _selfRolesByGuildAndId = new ConcurrentDictionary<ulong, ConcurrentDictionary<ulong, SelfRole>>();
         LoadSelfRoleData();
@@ -31,8 +28,6 @@ public class SelfRolesProvider
 
     private void LoadSelfRoleData()
     {
-        var transaction = _sentry.StartNewConfiguredTransaction("StartUp", nameof(LoadSelfRoleData));
-
         _logger.LogInformation("Loading data from database: SelfRoles");
 
         SelfRole[] allRoles;
@@ -41,12 +36,7 @@ public class SelfRolesProvider
             allRoles = db.SelfRoles.AsNoTracking().Include(sr => sr.Guild).ToArray();
         }
 
-        foreach (var role in allRoles)
-        {
-            AddRoleToInternalCache(role.Guild.DiscordId, role);
-        }
-
-        transaction.Finish();
+        foreach (var role in allRoles) AddRoleToInternalCache(role.Guild.DiscordId, role);
     }
 
     private void AddRoleToInternalCache(ulong guildId, SelfRole role)
@@ -57,23 +47,14 @@ public class SelfRolesProvider
 
     private void RemoveRoleFromInternalCache(ulong guildId, ulong roleId)
     {
-        if (_selfRolesByGuildAndId.TryGetValue(guildId, out var roles))
-        {
-            roles.TryRemove(roleId, out _);
-        }
+        if (_selfRolesByGuildAndId.TryGetValue(guildId, out var roles)) roles.TryRemove(roleId, out _);
     }
 
-    public bool BotKnowsRole(ulong guildId, ulong roleId)
-    {
-        return _selfRolesByGuildAndId.TryGetValue(guildId, out var roles) ? roles.ContainsKey(roleId) : false;
-    }
+    public bool BotKnowsRole(ulong guildId, ulong roleId) => _selfRolesByGuildAndId.TryGetValue(guildId, out var roles) ? roles.ContainsKey(roleId) : false;
 
     public bool TryGetRole(ulong guildId, ulong roleId, out SelfRole role)
     {
-        if (_selfRolesByGuildAndId.TryGetValue(guildId, out var roles))
-        {
-            return roles.TryGetValue(roleId, out role);
-        }
+        if (_selfRolesByGuildAndId.TryGetValue(guildId, out var roles)) return roles.TryGetValue(roleId, out role);
         role = null;
         return false;
     }
@@ -81,7 +62,7 @@ public class SelfRolesProvider
     public bool TryGetRole(ulong guildId, string roleName, out SelfRole role)
     {
         role = _selfRolesByGuildAndId.TryGetValue(guildId, out var guildRoles)
-            ? guildRoles.Values.FirstOrDefault(x => x.RoleName.Equals(roleName, System.StringComparison.OrdinalIgnoreCase))
+            ? guildRoles.Values.FirstOrDefault(x => x.RoleName.Equals(roleName, StringComparison.OrdinalIgnoreCase))
             : null;
         return role != null;
     }
@@ -93,38 +74,23 @@ public class SelfRolesProvider
             roles = indexedRoles.Values.ToList();
             return true;
         }
+
         roles = new List<SelfRole>();
         return false;
     }
 
     public async Task AddRole(ulong guildId, SelfRole role)
     {
-        var transaction = _sentry.StartSpanOnCurrentTransaction(nameof(AddRole));
-
-        if (!BotKnowsRole(guildId, role.DiscordRoleId))
-        {
-            await InsertSelfRole(guildId, role);
-        }
-
-        transaction.Finish();
+        if (!BotKnowsRole(guildId, role.DiscordRoleId)) await InsertSelfRole(guildId, role);
     }
 
     public async Task RemoveRole(ulong guildId, ulong roleId)
     {
-        var transaction = _sentry.StartSpanOnCurrentTransaction(nameof(RemoveRole));
-
-        if (TryGetRole(guildId, roleId, out var role))
-        {
-            await DeleteSelfRole(guildId, role);
-        }
-
-        transaction.Finish();
+        if (TryGetRole(guildId, roleId, out var role)) await DeleteSelfRole(guildId, role);
     }
 
     private async Task InsertSelfRole(ulong guildId, SelfRole role)
     {
-        var transaction = _sentry.StartSpanOnCurrentTransaction(nameof(InsertSelfRole));
-
         _logger.LogInformation("Writing a new Self Role {RoleName} for Guild {GuildId} to the database", role.RoleName, guildId);
 
         int writtenCount;
@@ -135,21 +101,13 @@ public class SelfRolesProvider
         }
 
         if (writtenCount > 0)
-        {
             AddRoleToInternalCache(guildId, role);
-        }
         else
-        {
             _logger.LogError("Writing Self Role {RoleName} for Guild {GuildId} to the database inserted no entities. The internal cache was not changed", role.RoleName, guildId);
-        }
-
-        transaction.Finish();
     }
 
     private async Task DeleteSelfRole(ulong guildId, SelfRole role)
     {
-        var transaction = _sentry.StartSpanOnCurrentTransaction(nameof(DeleteSelfRole));
-
         _logger.LogInformation("Deleting Self Role {RoleName} for Guild {GuildId} from the database", role.RoleName, guildId);
 
         int writtenCount;
@@ -160,14 +118,8 @@ public class SelfRolesProvider
         }
 
         if (writtenCount > 0)
-        {
             RemoveRoleFromInternalCache(guildId, role.DiscordRoleId);
-        }
         else
-        {
             _logger.LogWarning("Deleting Self Role {RoleName} for Guild {GuildId} from the database deleted no entities. The internal cache was not changed", role.RoleName, guildId);
-        }
-
-        transaction.Finish();
     }
 }
